implemented ProductController

+++ buy_sale.database/BuySaleDbContext.cs
@@ -8,6 +8,7 @@ namespace buy_sale.database
 {
     public class BuySaleDbContext : DbContext
     {
+        private static bool instance = false;
         public DbSet<Product> Products { get; set; }
         public DbSet<SalesPoint> SalesPoints { get; set; }
         public DbSet<ProvidedProducts> ProvidedProducts { get; set; }
@@ -17,8 +18,12 @@ namespace buy_sale.database
 
         public BuySaleDbContext(DbContextOptions<BuySaleDbContext> options) : base(options)
         {
-            Database.EnsureDeleted();
-            Database.EnsureCreated();
+            if (!instance)
+            {
+                Database.EnsureDeleted();
+                Database.EnsureCreated();
+                instance = true;
+            }        
         }
 
         protected override void OnModelCreating(ModelBuilder modelBuilder)
@@ -32,36 +37,37 @@ namespace buy_sale.database
         {
             var products = new List<Product>()
             {
-                new Product { Name = "Мыло", Price = 0.25m },
-                new Product { Name = "Верёвка", Price = 5.25m },
-                new Product { Name = "Стул", Price = 10.25m }
+                new Product { Id = 1, Name = "Мыло", Price = 0.25m },
+                new Product { Id = 2, Name = "Верёвка", Price = 5.25m },
+                new Product { Id = 3, Name = "Стул", Price = 10.25m }
             };
             modelBuilder.Entity<Product>().HasData(products);
-
+            
             var salesPoints = new List<SalesPoint>()
             {
-                new SalesPoint {Name = "Не кисни - зависни"},
+                new SalesPoint {Id = 1, Name = "Не кисни - зависни"},
             };
             modelBuilder.Entity<SalesPoint>().HasData(salesPoints);
 
             var providedProducts = new List<ProvidedProducts>
             {
-                new ProvidedProducts { Id = 1, ProductId = products[0].Id, ProductQuantity = 6, SalesPointId = salesPoints[0].Id },
-                new ProvidedProducts { Id = 2, ProductId = products[1].Id, ProductQuantity = 3, SalesPointId = salesPoints[0].Id },
-                new ProvidedProducts { Id = 3, ProductId = products[2].Id, ProductQuantity = 12, SalesPointId = salesPoints[0].Id }
+                new ProvidedProducts { Id = 1, ProductId = 1, ProductQuantity = 6, SalesPointId = 1 },
+                new ProvidedProducts { Id = 2, ProductId = 2, ProductQuantity = 3, SalesPointId = 1 },
+                new ProvidedProducts { Id = 3, ProductId = 3, ProductQuantity = 12, SalesPointId = 1 }
             };
             modelBuilder.Entity<ProvidedProducts>().HasData(providedProducts);
 
             var buyers = new List<Buyer>
             {
-                new Buyer { Name = "Максим" },
-                new Buyer { Name = "Вадим" }
+                new Buyer { Id = 1, Name = "Максим" },
+                new Buyer { Id = 2, Name = "Вадим" }
             };
             modelBuilder.Entity<Buyer>().HasData(buyers);
 
             var sales = new List<Sale>
             {
                 new Sale {
+                    Id = 1,
                     Date = DateOnly.FromDateTime(DateTime.Now),
                     Time = TimeOnly.FromDateTime(DateTime.Now),
                     SalesPointId = salesPoints[0].Id,
